redfine notify_event and send_client_message

+++ gui/ctrl/splitter.cpp
@@ -91,7 +91,7 @@ namespace gui {
         if (new_x != pt.x()) {
 //          logging::debug() << "splitter::set_value old:" << pt.x() << " new:" << new_x;
           position(core::point(new_x, pt.y()), true);
-          super::notify_event(detail::SLIDER_MESSAGE, static_cast<long>(new_x - pt.x()));
+          super::notify_event(detail::SLIDER_MESSAGE, static_cast<int>(new_x - pt.x()));
         }
 
       }
@@ -155,7 +155,7 @@ namespace gui {
         if (new_y != pt.y()) {
 //          logging::debug() << "splitter::set_value old:" << pt.y() << " new:" << new_y;
           position(core::point(pt.x(), new_y), true);
-          super::notify_event(detail::SLIDER_MESSAGE, static_cast<long>(new_y - pt.y()));
+          super::notify_event(detail::SLIDER_MESSAGE, static_cast<int>(new_y - pt.y()));
         }
 
       }
